# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Founders Finance API
servers:
  - url: /api
    description: Base API path
...
security:
  - sessionCookie: []
paths:
...
  /owner-draws:
    get:
      operationId: listOwnerDraws
      tags: [owner-draws]
      summary: List owner draws
      responses:
        "200":
          description: List of owner draws
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OwnerDraw"
    post:
      operationId: createOwnerDraw
      tags: [owner-draws]
      summary: Record an owner draw and its balanced journal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOwnerDrawBody"
      responses:
        "201":
          description: Created owner draw
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerDraw"
...
  /reimbursements/{id}/mark-paid:
    post:
      operationId: markReimbursementPaid
      tags: [reimbursements]
      summary: Mark reimbursement as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkPaidBody"
      responses:
        "200":
          description: Updated reimbursement request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReimbursementRequest"
  /reimbursements/{id}/waive:
    post:
      operationId: waiveReimbursement
      tags: [reimbursements]
      summary: Waive reimbursement with a balanced write-off
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReimbursementResolutionBody"
      responses:
        "200":
          description: Waived reimbursement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReimbursementRequest"
  /reimbursements/{id}/convert-to-contribution:
    post:
      operationId: convertReimbursementToContribution
      tags: [reimbursements]
      summary: Convert a Personal-funded reimbursement into owner capital
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReimbursementResolutionBody"
      responses:
        "200":
          description: Converted reimbursement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReimbursementRequest"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ff_session
...
  schemas:
...
    MarkPaidBody:
      type: object
      properties:
        payment_date:
          type: string
          format: date
        memo:
          type: string
          nullable: true
...
    OwnerDraw:
      type: object
      properties:
        id:
          type: string
          format: uuid
        transaction_id:
          type: string
          format: uuid
          nullable: true
        entity_id:
          type: string
          format: uuid
        entity_display_name:
          type: string
          nullable: true
        entity_primary_color:
          type: string
          nullable: true
        amount:
          type: number
        memo:
          type: string
          nullable: true
        draw_date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
      required:
        - id
        - entity_id
        - amount
        - draw_date
        - created_at
    CreateOwnerDrawBody:
      type: object
      properties:
        entity_id:
          type: string
          format: uuid
        amount:
          type: number
          exclusiveMinimum: 0
        memo:
          type: string
          nullable: true
        draw_date:
          type: string
          format: date
      required:
        - entity_id
        - amount
        - draw_date
    ReimbursementRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        original_transaction_id:
          type: string
          format: uuid
          nullable: true
        owed_to_entity_id:
          type: string
          format: uuid
        owed_to_entity_name:
          type: string
          nullable: true
        owed_to_entity_color:
          type: string
          nullable: true
        owed_to_entity_short_code:
          type: string
          nullable: true
        owed_by_entity_id:
          type: string
          format: uuid
        owed_by_entity_name:
          type: string
          nullable: true
        owed_by_entity_color:
          type: string
          nullable: true
        amount:
          type: number
        status:
          type: string
          enum:
            - pending
            - partially_paid
            - paid
            - waived
            - converted
        memo:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - owed_to_entity_id
        - owed_by_entity_id
        - amount
        - status
        - created_at
        - updated_at
    ReimbursementResolutionBody:
      type: object
      properties:
        effective_date:
          type: string
          format: date
        memo:
          type: string
          minLength: 3
      required:
        - memo
...